# By Michael Herold (15) and others
# Via marcslavec (6) and others
* 'master' of github.com:Catrobat/Catty: (25 commits)
  fix protocol warning issue
  fix all parser/serializer tests
  fixes #419
  rename unused initialization method
  removed CBXMLLogger categories
  Remove compiler warnings for Release Build
  dealloc SpriteObject when Program is deallocated
  dealloc Program after opening downloaded program
  dealloc AlertView & Program after renaming program
  dealloc Program after Scene is stopped
  slider improvement for better handling
  delegate changes
  some improvements for paint tools
  duplicate image when copied Look is modified
  Copy SpriteObject IDE bugfix
  added CBMutableCopying protocol to Formula
  Merge branch 'master' into issue#308
  license changed to 2015
  Update README.md
  Update README.md
  ...

+++ src/Catty/Extension&Delegate&Protocol/Extensions/GDataXMLElement/GDataXMLElement+CustomExtensions.m
@@ -1,5 +1,5 @@
 /**
- *  Copyright (C) 2010-2014 The Catrobat Team
+ *  Copyright (C) 2010-2015 The Catrobat Team
  *  (http://developer.catrobat.org/credits)
  *
  *  This program is free software: you can redistribute it and/or modify
@@ -149,7 +149,7 @@
     [XMLError exceptionIfNil:name message:@"Given param xmlElement MUST NOT be nil!!"];
     if (context.currentPositionStack) {
         if (xPathIndex > 1) {
-            name = [name stringByAppendingFormat:@"[%lu]", xPathIndex];
+            name = [name stringByAppendingFormat:@"[%lu]", (unsigned long)xPathIndex];
         }
         [context.currentPositionStack pushXmlElementName:name];
         NSLog(@"+ [%@] added to stack", name);
